# app/backend/swaggerdoc.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manual Técnico do VolunteeringIPS",
        "version": "1.0.0",
        "title": "VolunteeringIPS Doc"
    },
    "basePath": "/api",
...
    "schemes": ["http"],
    "paths": {
        "/email/notify": {
            "post": {
                "summary": "Used to send an email to an user",
                "consumes": [
                    "application/json"
                ],
                "produces": ["application/json"],
                "tags": [
                    "Notifications"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "notification",
                        "description": "The notification to create",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Notification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email successfully sent"
                    },
                    "401": {
                        "description": "Email not sent due to unauthorization (invalid token)"
                    }
                }
            }
        },
...
        "/credentials/forgot": {
            "post": {
                "summary": "Used to send an email when the user forgets his account's password",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "forgot",
                        "description": "Email to reset the password and his notification after",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Forgot"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request successefully sent"
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "403": {
                        "description": "Non existent email"
                    }
                }
            }
        },
        "/credentials/reset": {
            "patch": {
                "summary": "Used to reset the password, when forgotten by the user",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "reset",
                        "description": "This will only work if a 'forgot request' was made before",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Reset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account's password successfully changed"
                    },
                    "400": {
                        "description": "Passwords aren't the same"
                    },
                    "404": {
                        "description": "Probably the Forgot Request was not made"
                    }
                }
            }
        },
        "/credentials/confirmAccount/{token}": {
            "get": {
                "summary": "Used to activate the account",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Generated token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account's activation done successfully"
                    },
                    "400": {
                        "description": "Invalid token!"
                    }
                }
            }
        },
        "/credentials/isLoggedIn": {
            "get": {
                "summary": "Used to see if the user is logged in",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "You're logged"
                    },
                    "401": {
                        "description": "Not logged or inactive account"
                    }
                }
            }
        },
        "/credentials/hasAccess": {
            "get": {
                "summary": "Used to see if the user has access to a certain page",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "You have access"
                    },
                    "401": {
                        "description": "You dont have access to this page"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Notification": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "auth": {
                    "type": "string"
                },
                "action": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "params": {
                    "type": "object"
                }
            }
        },
...
        "Forgot": {
            "type": "object",
            "required": [
                "email",
                "action"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "action": {
                    "type": "string"
                }
            }
        },
        "Reset": {
            "type": "object",
            "required": [
                "id",
                "password",
                "confirmPassword"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "confirmPassword": {
                    "type": "string"
                }
            }
        },
...
    }
}